+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/cache.php
@@ -22,6 +22,7 @@ use Symfony\Component\Cache\Adapter\MemcachedAdapter;
 use Symfony\Component\Cache\Adapter\PdoAdapter;
 use Symfony\Component\Cache\Adapter\ProxyAdapter;
 use Symfony\Component\Cache\Adapter\RedisAdapter;
+use Symfony\Component\Cache\Adapter\RedisTagAwareAdapter;
 use Symfony\Component\Cache\Adapter\TagAwareAdapter;
 use Symfony\Component\Cache\Marshaller\DefaultMarshaller;
 use Symfony\Component\HttpKernel\CacheClearer\Psr6CacheClearer;
@@ -133,6 +134,22 @@ return static function (ContainerConfigurator $container) {
             ])
 
         ->set('cache.adapter.redis', RedisAdapter::class)
+            ->abstract()
+            ->args([
+                abstract_arg('Redis connection service'),
+                '', // namespace
+                0, // default lifetime
+                service('cache.default_marshaller')->ignoreOnInvalid(),
+            ])
+            ->call('setLogger', [service('logger')->ignoreOnInvalid()])
+            ->tag('cache.pool', [
+                'provider' => 'cache.default_redis_provider',
+                'clearer' => 'cache.default_clearer',
+                'reset' => 'reset',
+            ])
+            ->tag('monolog.logger', ['channel' => 'cache'])
+
+        ->set('cache.adapter.redis_tag_aware', RedisTagAwareAdapter::class)
             ->abstract()
             ->args([
                 abstract_arg('Redis connection service'),